Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/svg/animations/animate-fill-freeze-with-repeatDur-expected.txt


FAIL Test for animation freeze when repeatDur is not a multiple of dur assert_approx_equals: expected 150 +/- 1 but got 50
